shuffle as_buildinfo, as_taginfo, and as_targetinfo into the main kojismokydingo package alongside the bulk loading utils
